==> Rscripts/02-species_records/02-state_and_territory_fields.R <==


# library --------------------------------------------------------
  library(tidyverse)
  library(data.table)
  
  rm(list = ls())
  
# APC native/nonnative (naturalised) tags -------------------------------------
# Australian plant census names (APC)                                              
  apc <- read.csv("Data files/ALA/supplied data/nslsimplename-2015-06-25-2021.csv", strip.white = TRUE)
  glimpse(apc)
  
# include rank of species who have APC names
  dat <- filter(apc, rank == "Species" & classifications %in% c("[APC, APNI]", "[APNI, APC]"))
  
  a <- dat$apc_distribution
  b <- strsplit(as.character(a), split = ",")
  n_col <- unlist(lapply(b, function(x) length(x)))
  
  c <- matrix("", nrow = nrow(dat), ncol = max(n_col))
  for(i in 1:length(a)) if(n_col[i] > 0) c[i, 1:n_col[i]] <- b[[i]]
  
# remove spaces before and after tags names (e.g. "_NSW" and "NSW_")
# function to trim leading and trailing white space
  trimm <- function( x ) {
    gsub("(^[[:space:]]+|[[:space:]]+$)", "", x)
  }  
  c <- apply(c, c(1,2), function(x) trimm(x))
  table(c)
  length(table(c))
  
  write.csv(table(c), "Data files/ALA/cleaning steps/distribution fields beta.csv")
  
# create distribution (state) fields for ALA records--------------------------------
# notes ------------------------------------------------------------------------------
# I manually updated the list we just generated by individually tagging each unique 'field'/ state tag as naturalised or native, in their own columns
# -----------------------------------------------------------------------------------
# updated distribution fields
  distr <- read.csv("Data files/ALA/cleaning steps/distribution fields updated.csv", header = T) %>%
    mutate(., native = ifelse(is.na(native) == T, 0, 1),
           naturalised = ifelse(is.na(naturalised) == T, 0, 1))
  
  names(distr) <- c("id", "field", "native", "naturalised", "freq")  
  head(distr)
  
# create vectors to speed things up
  field <- distr$field
  naturalised <- distr$naturalised
  native <- distr$native
  
# identiy whether each species is listed as native or naturalised from the fields
  exo <- rowSums(apply(c, c(1, 2), function(x) naturalised[which(field == x)]))
  nat <- rowSums(apply(c, c(1, 2), function(x) native[which(field == x)]))
  
# now combine these to determine exclusively naturalised species (i.e. no native entry)
  exo_nat <- ifelse(nat == 0 & exo > 0, 1, 0) 
  
# number of naturalised species
  sum(exo_nat) # 2,952 exotics
  
# native species that have naturalised elsewhere in Australia
  native_nat <- ifelse(nat == 1 & exo > 0, 1, 0) 
  sum(native_nat) # 150 natives that have adventured outwards
  
# list each species as native/naturalised (0/1)
  dat.final <- mutate(dat, naturalised = exo_nat, native_naturalised = native_nat) %>%
    dplyr:::select(name, naturalised, native_naturalised, apc_name, apc_familia, familia, author, authority, genus, species) %>%
    arrange(familia, name) # so naturalised = exotic and native_naturalised = aventurous native; we assume all else are native
  
# output
  write.csv(dat.final, "Data files/ALA/cleaning steps/step 1 - ALA names & status.csv")
  
# -------------------------------------------------------------------------------
